requirements.txt
openai>=1.3
datasets
python-dotenv
pandas